+++ src/Presentation/Nop.Web/Controllers/CountryController.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using System.Linq;
-using System.Web.Mvc;
-using Nop.Core;
-using Nop.Core.Caching;
-using Nop.Services.Directory;
-using Nop.Services.Localization;
+﻿using System.Web.Mvc;
+using Nop.Web.Factories;
 using Nop.Web.Framework;
-using Nop.Web.Infrastructure.Cache;
 
 namespace Nop.Web.Controllers
 {
@@ -14,27 +8,15 @@ namespace Nop.Web.Controllers
 	{
 		#region Fields
 
-        private readonly ICountryService _countryService;
-        private readonly IStateProvinceService _stateProvinceService;
-        private readonly ILocalizationService _localizationService;
-        private readonly IWorkContext _workContext;
-        private readonly ICacheManager _cacheManager;
+        private readonly ICountryModelFactory _countryModelFactory;
 
 	    #endregion
 
 		#region Constructors
 
-        public CountryController(ICountryService countryService, 
-            IStateProvinceService stateProvinceService, 
-            ILocalizationService localizationService, 
-            IWorkContext workContext,
-            ICacheManager cacheManager)
+        public CountryController(ICountryModelFactory countryModelFactory)
 		{
-            this._countryService = countryService;
-            this._stateProvinceService = stateProvinceService;
-            this._localizationService = localizationService;
-            this._workContext = workContext;
-            this._cacheManager = cacheManager;
+            this._countryModelFactory = countryModelFactory;
 		}
 
         #endregion
@@ -46,54 +28,8 @@ namespace Nop.Web.Controllers
         [AcceptVerbs(HttpVerbs.Get)]
         public ActionResult GetStatesByCountryId(string countryId, bool addSelectStateItem)
         {
-            //this action method gets called via an ajax request
-            if (String.IsNullOrEmpty(countryId))
-                throw new ArgumentNullException("countryId");
-
-            string cacheKey = string.Format(ModelCacheEventConsumer.STATEPROVINCES_BY_COUNTRY_MODEL_KEY, countryId, addSelectStateItem, _workContext.WorkingLanguage.Id);
-            var cacheModel = _cacheManager.Get(cacheKey, () =>
-            {
-                var country = _countryService.GetCountryById(Convert.ToInt32(countryId));
-                var states = _stateProvinceService.GetStateProvincesByCountryId(country != null ? country.Id : 0, _workContext.WorkingLanguage.Id).ToList();
-                var result = (from s in states
-                              select new { id = s.Id, name = s.GetLocalized(x => x.Name) })
-                              .ToList();
-
-
-                if (country == null)
-                {
-                    //country is not selected ("choose country" item)
-                    if (addSelectStateItem)
-                    {
-                        result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.SelectState") });
-                    }
-                    else
-                    {
-                        result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.OtherNonUS") });
-                    }
-                }
-                else
-                {
-                    //some country is selected
-                    if (!result.Any())
-                    {
-                        //country does not have states
-                        result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.OtherNonUS") });
-                    }
-                    else
-                    {
-                        //country has some states
-                        if (addSelectStateItem)
-                        {
-                            result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.SelectState") });
-                        }
-                    }
-                }
-
-                return result;
-            });
-            
-            return Json(cacheModel, JsonRequestBehavior.AllowGet);
+            var model = _countryModelFactory.GetStatesByCountryId(countryId, addSelectStateItem);
+            return Json(model, JsonRequestBehavior.AllowGet);
         }
 
         #endregion
