+++ src/Presentation/Nop.Web/Factories/CountryModelFactory.cs
@@ -1,0 +1,94 @@
+﻿using System;
+using System.Linq;
+using Nop.Core;
+using Nop.Core.Caching;
+using Nop.Services.Directory;
+using Nop.Services.Localization;
+using Nop.Web.Infrastructure.Cache;
+
+namespace Nop.Web.Factories
+{
+    public partial class CountryModelFactory : ICountryModelFactory
+    {
+		#region Fields
+
+        private readonly ICountryService _countryService;
+        private readonly IStateProvinceService _stateProvinceService;
+        private readonly ILocalizationService _localizationService;
+        private readonly IWorkContext _workContext;
+        private readonly ICacheManager _cacheManager;
+
+	    #endregion
+
+		#region Constructors
+
+        public CountryModelFactory(ICountryService countryService, 
+            IStateProvinceService stateProvinceService, 
+            ILocalizationService localizationService, 
+            IWorkContext workContext,
+            ICacheManager cacheManager)
+		{
+            this._countryService = countryService;
+            this._stateProvinceService = stateProvinceService;
+            this._localizationService = localizationService;
+            this._workContext = workContext;
+            this._cacheManager = cacheManager;
+		}
+
+        #endregion
+
+        #region Methods
+
+        public virtual dynamic GetStatesByCountryId(string countryId, bool addSelectStateItem)
+        {
+            if (String.IsNullOrEmpty(countryId))
+                throw new ArgumentNullException("countryId");
+
+            string cacheKey = string.Format(ModelCacheEventConsumer.STATEPROVINCES_BY_COUNTRY_MODEL_KEY, countryId, addSelectStateItem, _workContext.WorkingLanguage.Id);
+            var cachedModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var country = _countryService.GetCountryById(Convert.ToInt32(countryId));
+                var states = _stateProvinceService.GetStateProvincesByCountryId(country != null ? country.Id : 0, _workContext.WorkingLanguage.Id).ToList();
+                var result = (from s in states
+                              select new { id = s.Id, name = s.GetLocalized(x => x.Name) })
+                              .ToList();
+
+
+                if (country == null)
+                {
+                    //country is not selected ("choose country" item)
+                    if (addSelectStateItem)
+                    {
+                        result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.SelectState") });
+                    }
+                    else
+                    {
+                        result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.OtherNonUS") });
+                    }
+                }
+                else
+                {
+                    //some country is selected
+                    if (!result.Any())
+                    {
+                        //country does not have states
+                        result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.OtherNonUS") });
+                    }
+                    else
+                    {
+                        //country has some states
+                        if (addSelectStateItem)
+                        {
+                            result.Insert(0, new { id = 0, name = _localizationService.GetResource("Address.SelectState") });
+                        }
+                    }
+                }
+
+                return result;
+            });
+            return cachedModel;
+        }
+
+        #endregion
+    }
+}
