+++ src/Presentation/Nop.Web/Factories/ICountryModelFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace Nop.Web.Factories
+{
+    public partial interface ICountryModelFactory
+    {
+        dynamic GetStatesByCountryId(string countryId, bool addSelectStateItem);
+    }
+}
