+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -269,6 +269,8 @@
     <Compile Include="Extensions\HtmlExtensions.cs" />
     <Compile Include="Extensions\ControllerExtensions.cs" />
     <Compile Include="Extensions\MappingExtensions.cs" />
+    <Compile Include="Factories\ICountryModelFactory.cs" />
+    <Compile Include="Factories\CountryModelFactory.cs" />
     <Compile Include="Factories\IForumModelFactory.cs" />
     <Compile Include="Factories\ForumModelFactory.cs" />
     <Compile Include="Factories\IBlogModelFactory.cs" />
